+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 * Restructure metadata reader output.
 * Generalize file naming logic.
 * Create list_tables and list_table_versions functions.
+* Transform year-only time index level to integer dtype.
 
 
 0.1.0 (2019-04-09)